backend/requirements.txt
fastapi[standard]
uvicorn
pydantic
itsdangerous
sqlalchemy
pymysql
python-dotenv